splitting up foreach test file; got rid of incorrect move but unbroke execution_policy
