+++ tests/unit/parallel/algorithms/foreach.cpp
@@ -5,68 +5,10 @@
 
 #include <hpx/hpx_init.hpp>
 #include <hpx/hpx.hpp>
-#include <hpx/include/parallel_for_each.hpp>
-#include <hpx/util/lightweight_test.hpp>
 
-#include <boost/range/functions.hpp>
-
-#include "test_utils.hpp"
+#include "foreach_tests.hpp"
 
 ///////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_for_each(ExPolicy && policy, IteratorTag)
-{
-    BOOST_STATIC_ASSERT(hpx::parallel::is_execution_policy<ExPolicy>::value);
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(boost::begin(c), boost::end(c), std::rand());
-
-    hpx::parallel::for_each(std::forward<ExPolicy>(policy),
-        iterator(boost::begin(c)), iterator(boost::end(c)),
-        [](std::size_t& v) {
-            v = 42;
-        });
-
-    // verify values
-    std::size_t count = 0;
-    std::for_each(boost::begin(c), boost::end(c),
-        [&count](std::size_t v) -> void {
-            HPX_TEST_EQ(v, std::size_t(42));
-            ++count;
-        });
-    HPX_TEST_EQ(count, c.size());
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_for_each_async(ExPolicy && p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(boost::begin(c), boost::end(c), std::rand());
-
-    hpx::future<void> f =
-        hpx::parallel::for_each(std::forward<ExPolicy>(p),
-            iterator(boost::begin(c)), iterator(boost::end(c)),
-            [](std::size_t& v) {
-                v = 42;
-            });
-    f.wait();
-
-    // verify values
-    std::size_t count = 0;
-    std::for_each(boost::begin(c), boost::end(c),
-        [&count](std::size_t v) -> void {
-            HPX_TEST_EQ(v, std::size_t(42));
-            ++count;
-        });
-    HPX_TEST_EQ(count, c.size());
-}
-
 template <typename IteratorTag>
 void test_for_each()
 {
@@ -87,30 +29,6 @@ void test_for_each()
     test_for_each(execution_policy(par(task)), IteratorTag());
 }
 
-// template <typename IteratorTag>
-// void test_for_each_exec()
-// {
-//     using namespace hpx::parallel;
-//
-//     {
-//         hpx::threads::executors::local_priority_queue_executor exec;
-//         test_for_each(par(exec), IteratorTag());
-//     }
-//     {
-//         hpx::threads::executors::local_priority_queue_executor exec;
-//         test_for_each(task(exec), IteratorTag());
-//     }
-//
-//     {
-//         hpx::threads::executors::local_priority_queue_executor exec;
-//         test_for_each(execution_policy(par(exec)), IteratorTag());
-//     }
-//     {
-//         hpx::threads::executors::local_priority_queue_executor exec;
-//         test_for_each(execution_policy(task(exec)), IteratorTag());
-//     }
-// }
-
 void for_each_test()
 {
     test_for_each<std::random_access_iterator_tag>();
@@ -123,69 +41,6 @@ void for_each_test()
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_for_each_exception(ExPolicy && policy, IteratorTag)
-{
-    BOOST_STATIC_ASSERT(hpx::parallel::is_execution_policy<ExPolicy>::value);
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(boost::begin(c), boost::end(c), std::rand());
-
-    bool caught_exception = false;
-    try {
-        hpx::parallel::for_each(std::forward<ExPolicy>(policy),
-            iterator(boost::begin(c)), iterator(boost::end(c)),
-            [](std::size_t& v) { throw std::runtime_error("test"); });
-
-        HPX_TEST(false);
-    }
-    catch(hpx::exception_list const& e) {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(policy, e);
-    }
-    catch(...) {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_for_each_exception_async(ExPolicy && p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(boost::begin(c), boost::end(c), std::rand());
-
-    bool caught_exception = false;
-    bool returned_from_algorithm = false;
-    try {
-        hpx::future<void> f =
-            hpx::parallel::for_each(std::forward<ExPolicy>(p),
-                iterator(boost::begin(c)), iterator(boost::end(c)),
-                [](std::size_t& v) { throw std::runtime_error("test"); });
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch(hpx::exception_list const& e) {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(p, e);
-    }
-    catch(...) {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-    HPX_TEST(returned_from_algorithm);
-}
-
 template <typename IteratorTag>
 void test_for_each_exception()
 {
@@ -214,67 +69,6 @@ void for_each_exception_test()
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_for_each_bad_alloc(ExPolicy && policy, IteratorTag)
-{
-    BOOST_STATIC_ASSERT(hpx::parallel::is_execution_policy<ExPolicy>::value);
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(boost::begin(c), boost::end(c), std::rand());
-
-    bool caught_exception = false;
-    try {
-        hpx::parallel::for_each(std::forward<ExPolicy>(policy),
-            iterator(boost::begin(c)), iterator(boost::end(c)),
-            [](std::size_t& v) { throw std::bad_alloc(); });
-
-        HPX_TEST(false);
-    }
-    catch(std::bad_alloc const&) {
-        caught_exception = true;
-    }
-    catch(...) {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_for_each_bad_alloc_async(ExPolicy && p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(boost::begin(c), boost::end(c), std::rand());
-
-    bool caught_exception = false;
-    bool returned_from_algorithm = false;
-    try {
-        hpx::future<void> f =
-            hpx::parallel::for_each(std::forward<ExPolicy>(p),
-                iterator(boost::begin(c)), iterator(boost::end(c)),
-                [](std::size_t& v) { throw std::bad_alloc(); });
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch(std::bad_alloc const&) {
-        caught_exception = true;
-    }
-    catch(...) {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-    HPX_TEST(returned_from_algorithm);
-}
-
 template <typename IteratorTag>
 void test_for_each_bad_alloc()
 {
@@ -302,50 +96,6 @@ void for_each_bad_alloc_test()
     test_for_each_bad_alloc<std::input_iterator_tag>();
 }
 
-////////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy>
-void test_executors(ExPolicy && policy)
-{
-    using iterator_tag = std::random_access_iterator_tag;
-
-    test_for_each(std::forward<ExPolicy>(policy), iterator_tag());
-    test_for_each_exception(std::forward<ExPolicy>(policy), iterator_tag());
-    test_for_each_bad_alloc(std::forward<ExPolicy>(policy), iterator_tag());
-}
-
-template <typename ExPolicy>
-void test_executors_async(ExPolicy && p)
-{
-    using iterator_tag = std::random_access_iterator_tag;
-
-    test_for_each_async(std::forward<ExPolicy>(p), iterator_tag());
-    test_for_each_exception_async(
-        std::forward<ExPolicy>(p), iterator_tag());
-    test_for_each_bad_alloc_async(
-        std::forward<ExPolicy>(p), iterator_tag());
-}
-
-void for_each_executors_test()
-{
-    using namespace hpx::parallel;
-    {
-        parallel_executor exec;
-
-        test_executors(par.on(exec));
-        test_executors_async(par(task).on(exec));
-    }
-
-    {
-        sequential_executor exec;
-
-        test_executors(seq.on(exec));
-        test_executors_async(seq(task).on(exec));
-
-        test_executors(par.on(exec));
-        test_executors_async(par(task).on(exec));
-    }
-}
-
 ///////////////////////////////////////////////////////////////////////////////
 int hpx_main(boost::program_options::variables_map& vm)
 {
@@ -359,7 +109,6 @@ int hpx_main(boost::program_options::variables_map& vm)
     for_each_test();
     for_each_exception_test();
     for_each_bad_alloc_test();
-    for_each_executors_test();
     return hpx::finalize();
 }
